Desenvolvendo a logica para inserir sessão a pauta

+++ src/main/java/com/assembleia/votacao/domain/Voto.java
@@ -20,6 +20,6 @@ public class Voto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int idAssociado;
-    private int idPauta;
+    private Long idAssociado;
+    private Long idPauta;
     }
